Fix signature of indicator.

+++ src/indicator.rs
@@ -16,12 +16,12 @@
 /// ```
 ///
 /// This may be what you need if you went to indicate an error on the invalid data.
-pub fn indicated_msg(doc: &str, mut pos: usize) -> String {
+pub fn indicated_msg(doc: &str, mut pos: u64) -> String {
     let mut show_line = String::new();
     for (line, str_line) in doc.split('\n').enumerate() {
         let full_line = str_line.len() + 1;
-        if full_line > pos {
-            let column = pos;
+        if full_line as u64 > pos {
+            let column = pos as usize;
             show_line = format!(
                 "({}:{})\n{}\n{}^",
                 line + 1,
@@ -31,7 +31,7 @@ pub fn indicated_msg(doc: &str, mut pos: usize) -> String {
             );
             break;
         } else {
-            pos -= full_line;
+            pos -= full_line as u64;
         }
     }
     show_line
